src/benchmarks_rad/requirements.txt
pandas==2.2.0
pika==1.3.2
rdkit==2023.9.5
rich==13.7.1
pyyaml==6.0.1
numpy==1.26.4
duckdb==1.2.2
pyarrow==19.0.1